# Reduced the garbage collection frequency in runtests.py.