@adamchainz 2021-03-17
This seems to have been less effective on CI.

[Previous PR, sqlite, Python 3.9](https://djangoci.com/job/pull-requests-bionic/database=sqlite3,label=bionic-pr,python=python3.9/10786/consoleFull): 783 seconds
[This PR, sqlite, Python 3.9](https://djangoci.com/job/pull-requests-bionic/database=sqlite3,label=bionic-pr,python=python3.9/10787/consoleFull): 761 seconds

But this might just be noise.