@adamchainz 2021-03-18
Of course. I mostly opened this to see how it went first.