@adamchainz 2021-03-18
I pushed a comment. I also did another benchmark with [hyperfine](https://github.com/sharkdp/hyperfine) on some other tests.

Before:

```
$ hyperfine './runtests.py view_tests'
Benchmark #1: ./runtests.py view_tests
  Time (mean ± σ):      4.803 s ±  0.207 s    [User: 4.557 s, System: 0.188 s]
  Range (min … max):    4.660 s …  5.373 s    10 runs
```

After:

```
$ hyperfine './runtests.py view_tests'
Benchmark #1: ./runtests.py view_tests
  Time (mean ± σ):      4.474 s ±  0.109 s    [User: 4.252 s, System: 0.185 s]
  Range (min … max):    4.376 s …  4.766 s    10 runs
```

4.474/4.803 = 0.932, so about 7% faster here.