@adamchainz 2021-03-19
Did another more thorough benchmark.

Before:

```
$ hyperfine './runtests.py absolute_url_overrides admin_autodiscover admin_changelist admin_checks admin_custom_urls admin_default_site admin_docs admin_filters admin_inlines admin_ordering admin_registration admin_scripts admin_utils admin_views admin_widgets'
Benchmark #1: ./runtests.py absolute_url_overrides admin_autodiscover admin_changelist admin_checks admin_custom_urls admin_default_site admin_docs admin_filters admin_inlines admin_ordering admin_registration admin_scripts admin_utils admin_views admin_widgets
  Time (mean ± σ):     106.603 s ±  3.243 s    [User: 95.358 s, System: 10.080 s]
  Range (min … max):   102.238 s … 109.942 s    10 runs
```

After:

```
$ hyperfine './runtests.py absolute_url_overrides admin_autodiscover admin_changelist admin_checks admin_custom_urls admin_default_site admin_docs admin_filters admin_inlines admin_ordering admin_registration admin_scripts admin_utils admin_views admin_widgets'
Benchmark #1: ./runtests.py absolute_url_overrides admin_autodiscover admin_changelist admin_checks admin_custom_urls admin_default_site admin_docs admin_filters admin_inlines admin_ordering admin_registration admin_scripts admin_utils admin_views admin_widgets
  Time (mean ± σ):     100.946 s ±  4.201 s    [User: 88.717 s, System: 10.593 s]
  Range (min … max):   90.755 s … 104.702 s    10 runs
```


100.9/106.6 = 0.947 , so about 5% faster. The "after" run had more variance though, maybe my computer was getting tired - if we compare just the fastest runs it's ~10% faster.