@felixxm 2021-04-09
@adamchainz Thanks for the proposition, however this caused a big performance regression on Windows, +10 minutes:
- ~11 m (without this patch),
- ~21 m (with this patch).

In particular `dispatch.tests.DispatcherTests.test_cached_garbaged_collected` is almost 10 minutes slower :chart_with_upwards_trend: .