@adamchainz 2021-04-09
Thanks for sptting. That's really concerning! Do you have metrics to see if the process had hit swap? And do you know if the windows box has a magnetic rather than solid state disk?

It could have been just that test got unlucky and needed to crawl through the swap. That particular test is triggering GC but it doesn't seem that different to others in the same test case - it just seems to be the first test to run in the test case and thus the first call to `gc.collect()` in the process.

I'm still playing around with this technique. Another thing I'll try now is reducing the frequency of garbage collection, rather than disabling it outright. The default CPython threshold frequency of 700 does seem to be impractically low - Michael Kennedy of talkpython found it could be triggered multiple times in just one normal view. I think that could help reduce the unnecessary gc work whilst preventing memory usage from ballooning into swap space.