@felixxm 2021-04-09
> Do you have metrics to see if the process had hit swap? And do you know if the windows box has a magnetic rather than solid state disk?

No, sorry :shrug: