@adamchainz 2021-04-09
Okay, I've just pushed a commit reducing the threshold. Locally it has again about a 5% improvement - let's see if we get any problems on windows on CI.


Before:

```
$ hyperfine './runtests.py view_tests'
Benchmark #1: ./runtests.py view_tests
  Time (mean ± σ):      5.028 s ±  0.099 s    [User: 4.784 s, System: 0.204 s]
  Range (min … max):    4.889 s …  5.180 s    10 runs
```

After:

```
$ hyperfine './runtests.py view_tests'
Benchmark #1: ./runtests.py view_tests
  Time (mean ± σ):      4.793 s ±  0.101 s    [User: 4.560 s, System: 0.203 s]
  Range (min … max):    4.586 s …  4.914 s    10 runs
```