@adamchainz 2021-04-09
Well it looks like the windows run was just fine this time, finishing in 801s.

And in fact ~10% faster than a [comparable run](https://djangoci.com/job/pull-requests-windows/database=sqlite3,label=windows,python=Python39/10994/) for another open PR, that finished in 884s.

`pr-mariadb/database=mysql,label=mariadb,python=python3.9`  also seems to have gone from 991s to 951s.